Route desktop broker services through capabilities
