+++ src/capabilities/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./registry";
 export * from "./factories";
+export * from "./service-capabilities";
