+++ src/renderers/electrobun/view/ai-host.ts
@@ -1,14 +1,22 @@
 import {
+  AI_RUNNER_CAPABILITY_ID,
+  type AiRunnerEvent,
+} from "../../../capabilities";
+import {
   getAiProviderDefinitions,
   __setDetectedProvidersForTests,
 } from "../../../plugins/builtin/ai/providers";
-import { setAiRunHost } from "../../../plugins/builtin/ai/runner";
-import { backendRequest, onAiChunk } from "./backend-rpc";
+import { AiRunCancelledError, setAiRunHost } from "../../../plugins/builtin/ai/runner";
+import { backendRequest, onCapabilityEvent } from "./backend-rpc";
 
 let nextRunId = 1;
 
 export async function installElectrobunAiHost(): Promise<void> {
-  const availability = await backendRequest<Record<string, boolean>>("ai.getProviderAvailability");
+  const availability = await backendRequest<Record<string, boolean>>("capability.invoke", {
+    capabilityId: AI_RUNNER_CAPABILITY_ID,
+    operationId: "getProviderAvailability",
+    payload: {},
+  });
   const providers = getAiProviderDefinitions().map((definition) => ({
     ...definition,
     available: availability[definition.id] ?? false,
@@ -17,30 +25,67 @@ export async function installElectrobunAiHost(): Promise<void> {
   __setDetectedProvidersForTests(providers);
   setAiRunHost({
     run({ provider, prompt, cwd, onChunk }) {
-      const runId = `ai-run:${nextRunId++}`;
-      const disposeChunks = onAiChunk(runId, ({ output }) => {
-        onChunk?.(output);
-      });
+      const subscriptionId = `ai-run:${nextRunId++}`;
       let disposed = false;
+      let settled = false;
+      let disposeMessages: () => void = () => {};
+      let resolveDone: (output: string) => void = () => {};
+      let rejectDone: (error: unknown) => void = () => {};
 
       const cleanup = () => {
         if (disposed) return;
         disposed = true;
-        disposeChunks();
+        disposeMessages();
+        void backendRequest("capability.unsubscribe", { subscriptionId }).catch(() => {});
       };
 
-      const done = backendRequest<string>("ai.run", {
-        runId,
-        providerId: provider.id,
-        prompt,
-        cwd,
-      }).finally(cleanup);
+      const settle = (callback: () => void) => {
+        if (settled) return;
+        settled = true;
+        cleanup();
+        callback();
+      };
+
+      const done = new Promise<string>((resolve, reject) => {
+        resolveDone = resolve;
+        rejectDone = reject;
+      });
+
+      disposeMessages = onCapabilityEvent(subscriptionId, (message) => {
+        const event = message.event as AiRunnerEvent;
+        switch (event.kind) {
+          case "chunk":
+            onChunk?.(event.output);
+            break;
+          case "done":
+            settle(() => resolveDone(event.output));
+            break;
+          case "cancelled":
+            settle(() => rejectDone(new AiRunCancelledError()));
+            break;
+          case "error":
+            settle(() => rejectDone(new Error(event.error)));
+            break;
+        }
+      });
+
+      void backendRequest("capability.subscribe", {
+        subscriptionId,
+        capabilityId: AI_RUNNER_CAPABILITY_ID,
+        operationId: "run",
+        payload: {
+          providerId: provider.id,
+          prompt,
+          cwd,
+        },
+      }).catch((error) => {
+        settle(() => rejectDone(error));
+      });
 
       return {
         done,
         cancel() {
-          cleanup();
-          void backendRequest("ai.cancel", { runId }).catch(() => {});
+          settle(() => rejectDone(new AiRunCancelledError()));
         },
       };
     },
