+++ src/renderers/electrobun/view/notes-files.ts
@@ -1,3 +1,4 @@
+import { NOTES_FILES_CAPABILITY_ID } from "../../../capabilities";
 import { backendRequest } from "./backend-rpc";
 
 export interface QuickNoteEntry {
@@ -9,37 +10,42 @@ export interface QuickNoteEntry {
 export class NotesFiles {
   constructor(private readonly dataDir: string) {}
 
+  private invoke<T>(operationId: string, payload: Record<string, unknown> = {}): Promise<T> {
+    return backendRequest<T>("capability.invoke", {
+      capabilityId: NOTES_FILES_CAPABILITY_ID,
+      operationId,
+      payload: {
+        dataDir: this.dataDir,
+        ...payload,
+      },
+    });
+  }
+
   async load(symbol: string): Promise<string> {
-    return backendRequest<string>("notes.load", {
-      dataDir: this.dataDir,
+    return this.invoke<string>("load", {
       symbol,
     });
   }
 
   async save(symbol: string, notes: string): Promise<void> {
-    await backendRequest("notes.save", {
-      dataDir: this.dataDir,
+    await this.invoke("save", {
       symbol,
       notes,
     });
   }
 
   async delete(symbol: string): Promise<void> {
-    await backendRequest("notes.delete", {
-      dataDir: this.dataDir,
+    await this.invoke("delete", {
       symbol,
     });
   }
 
   async loadQuickNotesIndex(): Promise<QuickNoteEntry[]> {
-    return backendRequest<QuickNoteEntry[]>("notes.loadQuickNotesIndex", {
-      dataDir: this.dataDir,
-    });
+    return this.invoke<QuickNoteEntry[]>("loadQuickNotesIndex");
   }
 
   async saveQuickNotesIndex(entries: QuickNoteEntry[]): Promise<void> {
-    await backendRequest("notes.saveQuickNotesIndex", {
-      dataDir: this.dataDir,
+    await this.invoke("saveQuickNotesIndex", {
       entries,
     });
   }
